delete and submit working

+++ bangazonAPI/views/order.py
@@ -122,18 +122,31 @@ class Orders(ViewSet):
         if request.method == "PUT":
             current_user = Customer.objects.get(user=request.auth.user)
 
-            try:
-                open_order = Order.objects.get(customer=current_user, paymenttype=None)
-                products_on_order = Product.objects.filter(cart__order=open_order)
-                product = Product.objects.get(pk=request.data["product_id"])
-                delete_me = OrderProduct.objects.filter(product=product, order=open_order)[0]
-                delete_me.delete()
+            if "payment_id" in request.data:
+                try:
+                    open_order = Order.objects.get(customer=current_user, paymenttype=None)
+                    payment_id = PaymentType.objects.get(pk=request.data["payment_id"])
+                    open_order.paymenttype = payment_id
+                    open_order.save()
+                    return Response({}, status=status.HTTP_204_NO_CONTENT)
 
-            except Order.DoesNotExist as ex:
-                return Response({'message': ex.args[0]}, status=status.HTTP_404_NOT_FOUND)
+                except Order.DoesNotExist as ex:
+                    return Response({'message': ex.args[0]}, status=status.HTTP_404_NOT_FOUND)
 
-            serializer = ProductSerializer(products_on_order, many=True, context={'request': request})
-            return Response(serializer.data)
+            if "product_id" in request.data:
+
+                try:
+                    open_order = Order.objects.get(customer=current_user, paymenttype=None)
+                    products_on_order = Product.objects.filter(cart__order=open_order)
+                    product = Product.objects.get(pk=request.data["product_id"])
+                    delete_me = OrderProduct.objects.filter(product=product, order=open_order)[0]
+                    delete_me.delete()
+
+                except Order.DoesNotExist as ex:
+                    return Response({'message': ex.args[0]}, status=status.HTTP_404_NOT_FOUND)
+
+                serializer = ProductSerializer(products_on_order, many=True, context={'request': request})
+                return Response(serializer.data)
 
 
     def list(self, request):
